% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.R
\docType{data}
\name{mplus.fontfamilies}
\alias{mplus.fontfamilies}
\title{M+ Font Families}
\format{list}
\usage{
mplus.fontfamilies
}
\description{
Convenient List of font families in M+ fonts.
You can use this to pass family name options to any one of the seven themes provided by this package.
}
\keyword{datasets}
