% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.R
\docType{data}
\name{mplus.fonttable}
\alias{mplus.fonttable}
\title{M+ Font Table}
\format{data.frame}
\usage{
mplus.fonttable
}
\description{
fonttable of M+ fonts. Not intended for general use.
}
\keyword{datasets}
